# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API lists for E-Rescues",
        "title": "Swagger UI for E-Rescues",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:19000",
    "basePath": "/api/v1",
    "paths": {
        "/judge/advice": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "input: ScoreStress, ScoreDepress, ScoreAnxiety =\u003e output: struct judge-advice\nWARNING: ALl the score must be greater than 1 and less than 10",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "judge-manager-apis"
                ],
                "summary": "Get judge-advice by score",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "score_Stress",
                        "name": "score_stress",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "score_depress",
                        "name": "score_depress",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "score_anxiety",
                        "name": "score_anxiety",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/judge/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "output: struct judges",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "judge-manager-apis"
                ],
                "summary": "Get judges",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/judge/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "input: judge model.Judge =\u003e output: status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "judge-manager-apis"
                ],
                "summary": "Create judges",
                "parameters": [
                    {
                        "description": "fill judge",
                        "name": "judge",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Judge"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/judge/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "input: judge's id =\u003e output: status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "judge-manager-apis"
                ],
                "summary": "Delete judge by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "judge's id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/judge/update": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "input: judge model.judge =\u003e output: status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "judge-manager-apis"
                ],
                "summary": "Update judges",
                "parameters": [
                    {
                        "description": "change judge",
                        "name": "judge",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Judge"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/judge/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "input: judge's id =\u003e output: struct judge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "judge-manager-apis"
                ],
                "summary": "Get judge by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "judge's id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Judge": {
            "type": "object",
            "properties": {
                "advice_anxiety": {
                    "type": "string"
                },
                "advice_depess": {
                    "type": "string"
                },
                "advice_stress": {
                    "description": "Lời khuyên tương ứng",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "score_anxiety_max": {
                    "type": "integer"
                },
                "score_anxiety_min": {
                    "type": "integer"
                },
                "score_depess_max": {
                    "type": "integer"
                },
                "score_depess_min": {
                    "type": "integer"
                },
                "score_stress_max": {
                    "description": "Phạm vi điểm số",
                    "type": "integer"
                },
                "score_stress_min": {
                    "type": "integer"
                }
            }
        },
...
        "model.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}